+++ tests/testthat/test-model_performance.rma.R
@@ -7,10 +7,25 @@ if (requiet("metafor")) {
     expect_null(mp$R2)
     expect_equal(mp$AIC, 28.40474, tolerance = 1e-3)
     expect_equal(mp$I2, 0.9222139, tolerance = 1e-3)
-    expect_equal(colnames(mp), c(
+    expect_identical(colnames(mp), c(
       "AIC", "BIC", "I2", "H2", "TAU2", "CochransQ",
       "p_CochransQ", "df_error", "Omnibus", "p_Omnibus"
     ))
+  })
+
+  test_that("check_outliers.rma", {
+    model <- rma(yi, vi, data = dat, slab = author, method = "REML")
+    out <- check_outliers(model)
+    expect_equal(
+      out,
+      c(
+        FALSE, FALSE, FALSE, TRUE, FALSE, FALSE, FALSE, TRUE,
+        FALSE, FALSE, FALSE, FALSE, FALSE
+      ),
+      ignore_attr = TRUE
+    )
+    expect_identical(attributes(out)$outlier_var, c("Hart & Sutherland", "TPT Madras"))
+    expect_snapshot(print(out))
   })
 
   test_that("model_performance.rma", {
@@ -19,7 +34,7 @@ if (requiet("metafor")) {
     expect_equal(mp$R2, 0.6463217, tolerance = 1e-3)
     expect_equal(mp$AIC, 24.21375, tolerance = 1e-3)
     expect_equal(mp$I2, 0.719778, tolerance = 1e-3)
-    expect_equal(colnames(mp), c(
+    expect_identical(colnames(mp), c(
       "AIC", "BIC", "I2", "H2", "TAU2", "CochransQ",
       "p_CochransQ", "df_error", "Omnibus", "p_Omnibus", "R2"
     ))